+++ sleepecg/heartbeats.py
@@ -102,8 +102,18 @@ def detect_heartbeats(ecg: np.ndarray, fs: float, backend: str = "c") -> np.ndar
         warnings.warn(f"Backend {backend!r} not available, using {fallback!r} instead.")
         backend = fallback
 
-    # For short signals, creating the bandpass filter makes up a large part of the total
-    # runtime. Therefore the filter is cached in a global variable.
+    # check for flat data at the beginning to avoid messing up detection thresholds
+    if len(ecg) < 2:
+        raise ValueError("ECG signal must have more than one sample.")
+    if ecg[1] == ecg[0]:
+        idx_nonflat = np.nonzero(ecg != ecg[0])[0]
+        if not len(idx_nonflat):
+            raise ValueError("ECG signal is flat. Please check your data.")
+        first_nonflat = idx_nonflat[0]  # first non-flat index
+    else:
+        first_nonflat = 0
+
+    # cache filter in global variable to speed up runtime (especially for short signals)
     try:
         sos = _sos_filters[fs]
     except KeyError:
@@ -116,8 +126,9 @@ def detect_heartbeats(ecg: np.ndarray, fs: float, backend: str = "c") -> np.ndar
         )
         _sos_filters[fs] = sos
 
-    # filtering bidirectionally removes filter delay
-    filtered_ecg = scipy.signal.sosfiltfilt(sos, ecg)
+    # filtering bidirectionally removes filter delay. Note that we start
+    # filtering the signal at the first non-flat index.
+    filtered_ecg = scipy.signal.sosfiltfilt(sos, ecg[first_nonflat:])
 
     # Set everything until the first zero-crossing to zero. For efficiency, only the first 2
     # seconds are checked.
@@ -148,7 +159,9 @@ def detect_heartbeats(ecg: np.ndarray, fs: float, backend: str = "c") -> np.ndar
         )
         beat_mask = _thresholding_py(filtered_ecg, integrated_ecg, fs)
 
-    return np.where(beat_mask)[0]
+    heartbeats = np.where(beat_mask)[0]
+    heartbeats += first_nonflat  # Add back the number of flat samples
+    return heartbeats
 
 
 class _CompareHeartbeatsResult(NamedTuple):